requirements.txt
# ── data collection ───────────────────────────────────────────────────────────
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# ── data modelling ────────────────────────────────────────────────────────────
pydantic==2.10.3

# ── embedding ─────────────────────────────────────────────────────────────────
sentence-transformers==5.3.0
numpy==2.1.3

# ── retrieval ─────────────────────────────────────────────────────────────────
faiss-cpu==1.13.2          # swap for faiss-gpu if running on GPU

# ── generation ────────────────────────────────────────────────────────────────
openai==2.30.0

# ── api ───────────────────────────────────────────────────────────────────────
fastapi==0.135.2
uvicorn==0.42.0

# ── config / utilities ────────────────────────────────────────────────────────
pyyaml==6.0.2
python-dotenv==1.1.0
